pyproject.toml
[project]
name = "transit-flow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.135.2",
    "firebase-admin>=7.3.0",
    "google-adk>=1.28.0",
    "google-auth>=2.49.1",
    "google-cloud-aiplatform>=1.143.0",
    "google-cloud-secret-manager>=2.27.0",
    "google-genai>=1.69.0",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "opik>=1.10.56",
    "pgvector>=0.4.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
]
